Attempting @JelleZijlstra's suggestion in https://github.com/python/cpython/issues/118761#issuecomment-2766771485. 

Running `python -Ximporttime -Sc  "import ast"` ten times with the current and proposed modules, I get 14.1ms for the current module and 1.48ms for the new, for a ~12ms speed up. I've left numbers out of the NEWS entry for now, as I've only tested on a Windows PC -- Linux may have different performance profiles.

<details>
<summary>Full Times</summary>

(using the last line of each output)

### Current

```text
import time: self [us] | cumulative | imported package
import time:      1277 |      13904 | ast
import time:      1363 |      14813 | ast
import time:      1343 |      14246 | ast
import time:      1330 |      14280 | ast
import time:      1387 |      13974 | ast
import time:      1373 |      13837 | ast
import time:      1345 |      14456 | ast
import time:      1349 |      14050 | ast
import time:      1322 |      14226 | ast
import time:      1246 |      13444 | ast
```

### Proposed

```text
import time:       455 |       1338 | ast
import time:       457 |       1387 | ast
import time:       476 |       1470 | ast
import time:       459 |       1474 | ast
import time:       493 |       1480 | ast
import time:       499 |       1458 | ast
import time:       448 |       1513 | ast
import time:       492 |       1501 | ast
import time:       487 |       1515 | ast
import time:       602 |       1688 | ast
```

</details> 

<!-- gh-issue-number: gh-118761 -->
* Issue: gh-118761
<!-- /gh-issue-number -->