@eli-schwartz 2025-04-01
```
$ wc -l /usr/lib/python3.{8,9,1*}/ast.py 
   549 /usr/lib/python3.8/ast.py
  1602 /usr/lib/python3.9/ast.py
  1709 /usr/lib/python3.10/ast.py
  1752 /usr/lib/python3.11/ast.py
  1840 /usr/lib/python3.12/ast.py
  1871 /usr/lib/python3.13/ast.py
  1777 /usr/lib/python3.14/ast.py
```

And with this PR we are back down to 671 lines. The massive jump is from when unparse was added.

This feels nice all on its own, to avoid so much code that is often not needed.