@AA-Turner 2025-04-01
> * I'm not sure we need to keep getattr support for the private attributes, like the Precedence enum and the various constants. I'd do that only if there's evidence they are being used in third-party code, especially widely used packages. Otherwise we just make our lives harder if we want to refactor the unparse code in the future.

Reverted.

> * An alternative approach could be to make `ast` a package, with the unparse code in `ast/_unparse.py` and the rest of the functionality in `ast/__init__.py`. Pro: Doesn't need a new top-level module; avoids the (probably theoretical) compatibility concern that someone could have their own top-level `_ast_unparse` module that interferes with the stdlib one. Con: Makes the directory structure more complex.

I'd suggest doing that in a follow-up PR, if desirable. I think for now this is the simplest approach. If it provides any reassurance, GitHub has [no results](https://github.com/search?q=Path%3A_ast_unparse.py&type=code) for the `_ast_unparse.py` file, so I think approaching 0% chance of breaking things in that regard.