@vstinner 2025-04-02
Benchmark: 13.1 ms => 18.8 ms (-5.7 ms).

Before:

```
$ hyperfine --warmup 1 "./python -c 'import ast'"
Benchmark 1: ./python -c 'import ast'
  Time (mean ± σ):      18.8 ms ±   0.6 ms    [User: 16.8 ms, System: 2.0 ms]
  Range (min … max):    17.9 ms …  20.3 ms    138 runs
```

After:

```
$ hyperfine --warmup 1 "./python -c 'import ast'"
Benchmark 1: ./python -c 'import ast'
  Time (mean ± σ):      13.1 ms ±   0.5 ms    [User: 11.1 ms, System: 2.0 ms]
  Range (min … max):    12.5 ms …  14.4 ms    206 runs
```