@vstinner 2025-04-02
> Sorry! Happy to open a follow-up PR for any of your suggestions?

Well, I don't know if it would better, but an alternative to adding `__dir__()` and `__getattr__()` functions is to add such wrapper function:

```py
def unparse(ast_obj):
    import _ast_unparse
    return _ast_unparse.unparse(ast_obj)
```