@danielhollas 2025-04-02
> Well, I don't know if it would better, but an alternative to adding __dir__() and __getattr__() functions is to add such wrapper function:

The `__dir__` implementation in particular seems a bit fragile. What happens if somebody introduces another stdlib dependency (besides sys) and forgets to exclude it in `__dir__` ?