@AA-Turner 2025-04-02
I think (untested) that the `__getattr__` approach is faster for the second access of `unparse` on, as it adds unparse to the module globals, whereas the proposed wrapper executes the (cached) `import _ast_unparse` each time. I'm happy to update `__dir__`, though.