@JelleZijlstra 2025-04-02
Alternative followup in #132025.